ranaming datappoints, trendMetrics and Trend, to metricPoint Metric, WindowedMetric based on team review
