+++ src/main/resources/resources/config/base.conf
@@ -5,14 +5,14 @@ kafka {
     num.stream.threads = 4
     commit.interval.ms = 3000
     auto.offset.reset = latest
-    timestamp.extractor = "com.expedia.www.haystack.datapoints.kstream.DataPointTimestampExtractor"
+    timestamp.extractor = "com.expedia.www.haystack.metricpoints.kstream.MetricPointTimestampExtractor"
   }
 
   producer {
-    topic = "aggregateddatapoints"
+    topic = "aggregatedmetricpoints"
   }
 
   consumer {
-    topic = "datapoints"
+    topic = "metricpoints"
   }
 }
